refactor(dapp): improve transactionToCreateAnAuction
